refactor: better code and comments
